requirements.txt
confluent-kafka>=2.6.0
python-dotenv>=1.0.0
streamlit>=1.40.0
requests>=2.32.0
